requirements.txt
pandas
numpy
requests
sqlalchemy
matplotlib
mysql-connector-python